Create README.md
In this video we'll look at the KivyMD DatePicker that allows you to pull up a calendar and pick a date, and then return that date to the screen. We can set a default day, or grab a range of dates or an individual date. We'll look at all those options in this video.